refactor(eventlog-live-otelcol): factor ~900 loc from main file into modules matching the hierarchy of the profiles processor
